Add files via upload
Find the average sale price for each sale condition.
Count the number of homes with each roof type for each neighborhood. Hint: Use the count of record IDs to count the number of homes.
Find the average number of above-grade bathrooms for each year built. Hint: To calculate the total number of bathrooms, add together the number of full baths, plus the number of half baths divided by two [full baths + (half baths/2)]. A house with two full bathrooms and one half bathroom, for example, has 2.5 total baths.
Find the maximum lot area for each combination of street type and building type.
Count the total number of homes with an alley for each neighborhood.
Find the average sale price for homes with each type of garage, along with the number of cars that the garage can hold.